% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{add_labels}
\alias{add_labels}
\title{Add task bar labels}
\usage{
add_labels(plt, task, label, position = "right")
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{task}{\code{data.frame}}

\item{label}{\code{string} decides which variable from task will be drawn as a label. Possible values are: 'people',
'task' and 'None'. If 'None', then labels are not added to the plot.}

\item{position}{\code{string} can be either 'left' or 'bottom', decides where label should be placed
referring to the task bar}
}
\value{
\code{ggplot object}
}
\description{
Adds geom_label layer with names given in \code{label} parameter. Labels are drawn next
to the task bars.
}
