% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{add_staging_areas}
\alias{add_staging_areas}
\title{Add coloring areas to denote stages}
\usage{
add_staging_areas(plt, task, cascade_staging)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{task}{\code{data.frame}}

\item{cascade_staging}{\code{string} decides how the staging areas should be created. If 'None', then
start and end of each stage are set as a start and end values of respectively first and last task within stage.
If 'on_finish' is provided, then each stage starts while last task of previous stage ends and lasts to the end
of the last task within stage.
On the other hand 'on_begin' option makes stage starting point with start of the first task in a stage and ending
point with start of first task in the next stage.}
}
\description{
Add coloring areas to denote stages
}
