% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{add_themes}
\alias{add_themes}
\title{Add theme to the plot}
\usage{
add_themes(plt, task, conf)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{task}{\code{data.frame}}

\item{conf}{\code{list} with styling configuration}
}
\description{
Adds styling to the plot with user defined configuration
}
