% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{adjust_path_data}
\alias{adjust_path_data}
\title{Recalculate control value}
\usage{
adjust_path_data(path_data, columns)
}
\arguments{
\item{path_data}{\code{data.frame}}

\item{columns}{\code{list}}
}
\description{
Recalculate control value
}
\details{
If path needs to be drawn within both stage and control group then
it is treated as drawing path within new control group which satisfy this need. This
function recalculates the control groups to fulfill this requirement. This can be seen if
some of tasks are in the same control group but in different stages. Then, after recalculation,
this tasks will have same control value within same stage but different within two different
stages. In other words, new, unique control value is submitted for pair (stage, control).
If path have to be drawn within one of stage of control, then the control is recalculated to fulfill this need.
}
