% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{assign_colors_to_fazes}
\alias{assign_colors_to_fazes}
\title{Assign color to tasks}
\usage{
assign_colors_to_fazes(task, color_by_fazes)
}
\arguments{
\item{task}{\code{data.frame}}

\item{color_by_fazes}{\code{logical} if \code{TRUE} then color assigned to each task is same for same fazes
, if \code{FALSE} then black color is assigned to every task.}
}
\value{
\code{list} of task with names of fazes
}
\description{
Assign color to tasks
}
\details{
Makes a named list of color codes, where names corresponds to unique fazes in \code{task}
}
\seealso{
\code{\link[gantt]{read_task}}
}
