% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{assign_colors_to_task_labels}
\alias{assign_colors_to_task_labels}
\title{Assign color to task labels}
\usage{
assign_colors_to_task_labels(task, color_by_stages)
}
\arguments{
\item{task}{\code{data.frame}}

\item{color_by_stages}{\code{logical} if \code{TRUE} then color assigned to each task label (y-axis) is same
within stage. If \code{FALSE} then black color is assigned to every task label.}
}
\value{
\code{list} of task with names of stages
}
\description{
Assign color to task labels
}
\details{
Makes a named list of color codes, where names corresponds to unique stages in \code{task}
}
\seealso{
\code{\link[gantt]{read_task}}
}
