% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_color_dict}
\alias{create_color_dict}
\title{stage color palette}
\usage{
create_color_dict(task)
}
\arguments{
\item{task}{\code{data.frame}}
}
\value{
\code{list} of color codes named by stages
}
\description{
stage color palette
}
\details{
Makes a named list of color codes, where names corresponds to unique stages in \code{task}
}
\seealso{
\code{\link[gantt]{read_task}}
}
