% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_legend}
\alias{create_legend}
\title{Create legend}
\usage{
create_legend(task)
}
\arguments{
\item{task}{\code{data.frame}}
}
\value{
\code{list} with colors, named by types of tasks
}
\description{
Create legend
}
\details{
Creates color list that can be used to manually scale colors in the plot. It is necessary because
bars and milestones are created as different geoms, but has to be considered in one legend.
}
\seealso{
\link[gantt]{draw_bars}
}
