% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_stage_task_list}
\alias{create_stage_task_list}
\title{Create list of tasks in each stages}
\usage{
create_stage_task_list(task)
}
\arguments{
\item{task}{\code{data.frame}}
}
\value{
\code{list} of task with names of tasks, named by stages
}
\description{
Create list of tasks in each stages
}
\details{
Creates a list in which each element is a \code{character} vector containing names of tasks
corresponding to every stage.
}
\seealso{
\code{\link[gantt]{read_task}}
}
