% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_summaries}
\alias{create_summaries}
\title{Create summary data frame}
\usage{
create_summaries(task, cascade_summaries = FALSE)
}
\arguments{
\item{task}{\code{data.frame}}

\item{cascade_summaries}{\code{logical} should stage just starts after previous ends (\code{TRUE}) or starts with first task
in the next stage (\code{FALSE}). Note, if there are stages that overlap then stage always starts with first task
(the \code{FALSE} equivalent).}
}
\value{
\code{data.frame} with stages summary
}
\description{
Creates data.frame which describes stages.
}
\details{
Each row has columns: 'stage', 'start', 'end', 'value' and 'type'.
Type is always set as 'summary', the value is calculated as maximum value of task
in a stage plus 0.3.
}
\seealso{
\code{\link[gantt]{read_task}}
}
