% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{delete_path_config}
\alias{delete_path_config}
\alias{delete_path_endings}
\title{Delete path endings}
\usage{
delete_path_endings(path_df)
}
\arguments{
\item{path_df}{\code{data.frame} with path points}
}
\description{
Delete path endings
}
\details{
To create paths in a way that there is a little horizontal line at the starting and ending point of task bars, path is created
as if tasks would start a day earlier and end one day later. In that manner there are unnecessary horizontal lines at path endings.
This function finds the starting and ending point of each visible path and shorten this path from both sides, so the endings are
not longer visible.
}
