% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{draw_arrows}
\alias{draw_arrows}
\title{Add arrows for stages}
\usage{
draw_arrows(plt, task, label_y_offset)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{task}{\code{data.frame}}

\item{label_y_offset}{\code{numeric} offset between arrow and its label}
}
\value{
\code{ggplot object}
}
\description{
Add arrows for stages
}
\details{
Adds plot layer with arrows describing each stage. Horizontal arrows are drawn under the earliest
task bar in each stage. Each starts in starting point of earliest task in the stage and ends in the ending
point of latest task in the the stage.
}
