% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{draw_bars}
\alias{draw_bars}
\title{Draw bars}
\usage{
draw_bars(plt, plot_task, bar_width, coloring_variable)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{plot_task}{\code{data.frame}}

\item{bar_width}{\code{numeric}}

\item{coloring_variable}{\code{string} either 'type' or 'stage'}
}
\description{
Draws bars for each task
}
