% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{draw_milestones}
\alias{draw_milestones}
\title{Draw milestone points}
\usage{
draw_milestones(plt, milestones_df)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{milestones_df}{\code{data.frame}, it is data.frame containing every task of milestone type}
}
\value{
\code{ggplot object}
}
\description{
Adds geom_point layer with points to corresponding milestones
}
