% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{draw_path}
\alias{draw_path}
\title{Draw path}
\usage{
draw_path(plt, task, is_stage_path, is_control_path)
}
\arguments{
\item{plt}{\code{ggplot object}}

\item{task}{\code{task}}

\item{is_stage_path}{\code{logical} if path will be drawn within stages}

\item{is_control_path}{\code{logical}  if path will be drawn within control groups}
}
\value{
\code{ggplot object}
}
\description{
Draw path
}
\details{
Draws path between bars on the plot. If only \code{is_stage_path} equals \code{TRUE} then
path is drawn between every bar within stages. If only \code{is_control_path} equals \code{TRUE} then
path is drawn between every bar within control groups. With both values equal \code{TRUE} then
path is drawn between every bar within stages and control groups.
Function draws a line by connecting points from \link[gantt]{critical_path} for valid stages and control groups.
Function groups path over control groups, previously modified with \link[gantt]{adjust_path_data}
which transforms control values in a way satisfying the given logic in \code{is_stage_path} and \code{is_control_path}.
}
