% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{middle_date}
\alias{middle_date}
\title{Get middle date}
\usage{
middle_date(start, end)
}
\arguments{
\item{start}{\code{date}}

\item{end}{\code{date}}
}
\description{
Calculates middle date between start and end.
}
