% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{read_task}
\alias{read_task}
\title{Read task}
\usage{
read_task(task_file = "task.csv", date_format = "\%d/\%m/\%y",
  delimiter = ";")
}
\arguments{
\item{task_file}{\code{string} path to csv file where tasks are specified}

\item{date_format}{\code{string} representing format of dates in task_file}

\item{delimiter}{\code{char} symbol which separates values in task_file}
}
\description{
Creates task dataframe from csv file
}
