% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_gantt_config}
\alias{create_gantt_config}
\title{Create configuration list}
\usage{
create_gantt_config(x_axis_text_size = 10, y_axis_text_size = 10,
  x_axis_label_position = "bottom", y_axis_label_position = "left",
  task_bar_color = "type", task_bar_width = 5,
  plot_title = "Gantt chart", plot_title_size = 15,
  stage_path = TRUE, control_path = TRUE, stage_arrows = TRUE,
  y_axis_color_by_stages = TRUE)
}
\arguments{
\item{x_axis_text_size}{\code{integer}}

\item{y_axis_text_size}{\code{integer}}

\item{x_axis_label_position}{\code{character} one of following: \code{top}, \code{bottom}}

\item{y_axis_label_position}{\code{character} one of following: \code{left}, \code{right}}

\item{task_bar_color}{decides which column from \code{type} and}

\item{task_bar_width}{\code{integer}}

\item{plot_title}{\code{character}}

\item{plot_title_size}{\code{integer}}

\item{stage_path}{\code{logical}, whether path will be drawn between tasks with same stage}

\item{control_path}{\code{logical}, whether path will be drawn between tasks with same control value}

\item{stage_arrows}{\code{logical}, whether arrows denoting stages have to be drawn}

\item{y_axis_color_by_stages}{\code{integer}}
}
\description{
Create configuration list
}
\details{
If \code{stage_path = TRUE} and \code{control_path = TRUE} then path is drawn within stages and within
control groups. \ If \code{stage_path = FALSE} and \code{control_path = FALSE} then no path is drawn
}
