% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_gantt_config}
\alias{create_gantt_config}
\title{Create configuration list}
\usage{
create_gantt_config(x_axis_text_size = 10, y_axis_text_size = 10,
  x_axis_label_position = "bottom", y_axis_label_position = "left",
  task_bar_color = "type", task_bar_width = 5,
  plot_title = "Gantt chart", plot_title_size = 15, faze_path = TRUE,
  control_path = TRUE, faze_arrows = TRUE,
  y_axis_color_by_fazes = TRUE)
}
\description{
Create configuration list
}
