% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{create_summaries}
\alias{create_summaries}
\title{Create summary data frame}
\usage{
create_summaries(task)
}
\arguments{
\item{task}{\code{data.frame}}
}
\value{
\code{data.frame} with fazes summary
}
\description{
Creates data.frame which describes fazes.
}
\details{
Each row has columns: 'faze', 'start', 'end', 'value' and 'type'.
Type is always set as 'summary', the value is calculated as maximum value of task
in a faze plus 0.3.
}
\seealso{
\code{\link[gantt]{read_task}}
}
