% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{mutate_task}
\alias{mutate_task}
\title{Mutate task}
\usage{
mutate_task(task)
}
\arguments{
\item{task}{\code{data.frame} with tasks specified}
}
\value{
Transformed \code{data.frame}
}
\description{
Transforms task data frame into ggplot friendly format
}
\details{
Transforms task \code{data.frame} in a way that each row from original \code{data.frame}
is doubled. Columns 'start' and 'end' are dropped and new column 'point' is added. Value of point
are missing value of 'start', 'end' columns.
}
\seealso{
\code{\link[gantt]{read_task}}
}
