% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{mutate_task}
\alias{mutate_task}
\title{Mutate task}
\usage{
mutate_task(task)
}
\arguments{
\item{task}{\code{data.frame} with tasks specified}
}
\description{
Transforms task data frame into ggplot friendly format
}
